Moves logic out from ContentRepositoryBase & into  a method on the Collection itself
