+++ src/Umbraco.Core/PropertyEditors/DataValueReferenceFactoryCollectionBuilder.cs
@@ -4,6 +4,6 @@ namespace Umbraco.Core.PropertyEditors
 {
     public class DataValueReferenceFactoryCollectionBuilder : OrderedCollectionBuilderBase<DataValueReferenceFactoryCollectionBuilder, DataValueReferenceFactoryCollection, IDataValueReferenceFactory>
     {
-        protected override DataValueReferenceFactoryCollectionBuilder This => this;         
+        protected override DataValueReferenceFactoryCollectionBuilder This => this;
     }
 }
